+++ rowbot2-data-importer-application/src/main/java/com/imatia/implatform/rowbot2/data/importer/application/services/externaldbs/importers/OracleImporter.java
@@ -42,7 +42,7 @@ public class OracleImporter extends AbstractJDBCImporter{
     }
 
     @Override
-    protected String getTableQuery(String qualifiedTableName, Long maxRowsToImport) {
+    protected String getTableQuery(String qualifiedTableName, Integer maxRowsToImport) {
         if(hasQueryLimit(maxRowsToImport)){
             return "SELECT * FROM " + qualifiedTableName + " OFFSET ? ROWS FETCH NEXT ? ROWS ONLY";
         }
